drbd_transport-lb_tcp: fix use of skb_seq_read()
skb_seq_read() returns the number of bytes available in the next
linear buffer. The caller has to take care that the buffer might be
only partially filled at this point. The code missed this so far.

Fixing this by limiting the number of bytes we look at. With this
change, I also introduce calling skb_abort_seq_read() at function exit
always (independently of the last return value of skb_seq_read()).

That allows us to drop the Linux-3.11 compat test and spatch rule.