drbd: set min_peer_disk_state to Inconsistent for PausedSyncT
In the same way that the disk state of a PausedSyncS node may be
Inconsistent, allow the peer disk state of a PausedSyncT node to be
Inconsistent. Otherwise the peer disk state is forced to Outdated, which
is asymmetric and causes P_OUT_OF_SYNC to be sent instead of P_DATA
unnecessarily.

This change mirrors:
e47dd7e1bab1 ("drbd: Do not assume that a L_PAUSED_SYNC_S has not a D_INCONSISTENT disk")