drbd: complete processing of P_RS_DEALLOCATED in receive_rs_deallocated()
When a resync target receives a P_RS_DEALLOCATED(_ID) packet, it must
set itself as out-of-sync relative to all other nodes. Precisely in
the same way as recv_resync_read() does it upon receiving a
P_RS_DATA_REPLY packet.
It also needs to update any peers to which the node is resync source
about the area that just became out of sync.

Failure to do so may cause awkward data inconsistencies between the
mirrors, which becomes noticeable as data corruption later on.