build: test for working coccinelle instead of version
We used a strict version check, which was for outdated, and also not
necessarily sufficient, as we saw with the two different 1.1.1 spatch versions
shipped in Ubuntu and Fedora.

Instead, we use a separate script that checks that spatch can actually parse
the patches, which is more robust than the version check.