build: test for working coccinelle instead of version
Second attempt, supersedes
288704a41 (build: test for working coccinelle instead of version, 2025-03-14)

Attempting to --parse-coccie each .cocci rule file individually won't work,
because some of our .cocci rule files depend on earlier ones in their series.

Also it may be overly strict, if the "problematic" rules are not even needed
for _this_ build.

Instead, first generate the aggregated rules file,
then syntax-check that result.

If our spatch is happy, try it.
If not, or if the actual spatch run fails,
still try to fall back to spatch-as-a-service (if enabled).