drbd: make "day0" bitmap slots reliable again
We use our "activiy log" to make the bitmap "crash safe".

In the "two node" case, we optimize away the activity log transactions on a
Secondary, if the only diskful peer is Primary, because the activity log
that is kept on that Primary is sufficient to be "crash safe".

This is no longer true if we have more than one bitmap slot, and connect a
"crashed" secondary (or a snapshot of a secondary, which basically looks the
same as "crashed" at the point in time of the snapshot) to a new node,
and expect an unassigned bitmap slot to cover all to-be synced blocks
(aka "day0" bitmap).

We must be more strict when enabling this optimization:
optimize away the activity log transactions on a Secondary
only if there is exactly one bitmap slot
(and there is only one diskful peer, and that is primary).

Fixes: 749ec5708 (drbd: Optimize AL updates in a two-node cluster, 2024-01-24)