drbd: let I/O drain when userspace freezes I/O on a DRBD device
The primary purpose of deliberately suspending I/O on a DRBD device is
to consistently snapshot the primary and secondary data.
I can not think of any use for suspending I/O without letting previous
write operations drain during the freeze operation. Therefore, I am
introducing the new behavior unconditionally.

In the request submit path, DRBD checks twice if the device has I/O
suspended. The first check is in inc_ap_bio(). If it fails, the code
sleeps in the caller's context. The second check is while holding the
state_rwlock. It is the relevant one. When it fails, it pushes the
request to the retry mechanism.

This patch also introduces a new list (resource->suspended_reqs) in
which postponed requests wait until the device resumes I/O.
Without this, the retry mechanism blocks the retry worker in
inc_ap_bio() when the device is I/O suspended.