drbd: set max_hw_zone_append_sectors to 0
Default stacking limits set this value to UINT_MAX, so it will end up as
UINT_MAX in our common limits. If we have a backing device, this will
get overwritten to 0 in the final blk_stack_limits, but if we are
diskless, this never happens.

There is a check in blk_validate_zoned_limits (called by
queue_limits_commit_update) that asserts that max_(hw_)zone_append_sectors
is 0. So this warning gets triggered in the diskless case.

Since we never support zoned devices in DRBD anyway, hard code this
value to 0 to avoid tripping the check.