drbd: send P_OUT_OF_SYNC after P_PEER_ACK in parallel
The main advantage of this is that it makes it easier to queue something
else to be done once all P_OUT_OF_SYNC have been sent. This will be used
for sending flush acks.

This change also fixes a potential race with del_connection() which
could have occurred as follows: The work is queued after the target
sender has been stopped. As a result, the work is never run. So
P_OUT_OF_SYNC is not sent for other connections after the one being
removed.