drbd: handle -EDESTADDRREQ in conn_connect
We observed a case where a resource was partially unregistered, but DRBD
was still trying to make a connection. The paths were already destroyed,
so the connection attempt failed, but DRBD continually tried to
establish a connection, resulting in flooding of the logs.

To make things slightly better, at least handle this case correctly in
conn_connect: if we don't have a destination address, we cannot possibly
make a connection, so quit trying there.