kmod spec: depmod: add a specific override for extra/drbd/
rpm macros put modules into /lib/modules/*/extra/.
depmod has a hardcoded default search path of "updates built-in".
Some distributions have a dist.conf that sets the depmod search path
to "updates extra built-in weak-modules", some even give precedence
to weak-modules over built-in. Others do as they do.
Some distributions come with the in-tree kernel module as "built-in".

If you install our kmod package, you want to use our modules.
Override the path to "extra/..." for the specific kernel version
this module was built against, and to "weak-updates/..." for all
other kernel versions.

Apparently it is not the most specific but the last match that wins.
List wildcard first, specific kernel release last.