drbd: drop references in w_update_peers() even if connection lost
If the peer only supports old protocol versions, or if we have lost the
connection, then we have "agreed_pro_version < 110". In this case,
w_update_peers() failed to free the work struct or drop the references
to the device and connection. As a result, it was not possible to rmmod
drbd.

Fix this by always cleaning up in w_update_peers().