drbd: only get kref if we are going to do a corresponding put
It was possible to get a kref in drbd_queue_send_out_of_sync() for a
connection that was already marked in the mask. Since only one put
occurred in drbd_send_oos_from() for this connection, we leaked a kref.

With this change there is a one-to-one correspondence:
kref_get()
<=> set the bit in send_oos_from_mask
<=> clear the bit in send_oos_from_mask (at some point)
<=> kref_put()

Fixes: 31f14b906312 ("drbd: send P_OUT_OF_SYNC after P_PEER_ACK in parallel")