drbd: track request "done" and kref separately
Until now the request kref was used to determine both when to remove the
request from transfer log, interval tree and activity log, and when to
destroy the request. However, the request was not always destroyed when
the kref reached 0. Sometimes it was used for queuing a peer ack. This
was surprising.

This change has subtle side-effects that should not impact the behavior
of DRBD. For instance, a request may be removed from the interval tree
and activity log while it is being iterated over in tl_walk() or debugfs
code paths.

This is preparation for a change to how a Primary node sends
P_OUT_OF_SYNC packets.