drbd: send P_OUT_OF_SYNC from Primary after updating bitmap
Instead of sending it in the stage for sending data.

This prevents problems when P_OUT_OF_SYNC is sent and then the peer
resyncs the blocks before the bits are set in the bitmap on the Primary.
The consequence is that the bits are left marked out-of-sync on the
Primary, even though the data is actually in sync. When a further write
to these blocks occurs, no P_OUT_OF_SYNC is sent because the Primary
believes that it is not necessary. This causes data corruption.

This problem can occur with the resync-without-replication feature. It
is very hard, or perhaps impossible, to trigger without using that
feature.

By sending P_OUT_OF_SYNC after setting the bits in the bitmap on the
Primary, we ensure that any bits which are set will be resynced.

It is possible as a result that blocks will be unnecessarily resynced.
This occurs when the blocks happen to be resynced after the local write
has occurred, but before the P_OUT_OF_SYNC is sent. This is harmless.