drbd_transport_tcp: drbd_transport_lb-tcp: fix default keepalive units
The module parameters drbd_keepidle and drbd_keepintvl are used in calls
to tcp_sock_set_keepidle() and tcp_sock_set_keepintvl(). These expect a
value in seconds. For instance, in the sources for Linux 6.15, the
implementation of each converts from seconds with "val * HZ".

In older kernels, the compatibility code replaces these lines with
"kernel_setsockopt(..., SOL_TCP, TCP_KEEPIDLE, ...)", respectively
TCP_KEEPINTVL. According to the man page tcp(7), this value is also in
seconds.

The effect of the incorrect units was that keepalive packets were only
sent every 23000s, reducing the effectiveness of the measure. Remove the
"*HZ" from our defaults so that they have the intended effect.

Fixes: 813bd569a ("drbd: TCP transports: set default keepalive time/interval/probes")