rdma: Modify rdma_conn_param.rnr_retry to 7
During the sync process, we found that the IB_WC_RNR_RETRY_EXC_ERR
error always occurred, and then the connection broke. Then, when the
connection is restored, this error occurs again, entering a loop.

When the IB_WC_RNR_RETRY_EXC_ERR error occurs, it indicates that there
is no rx_desc at the receive side currently.  Then we query that the
current rnr_retry value is 0, meaning there is no retry.

Set rnr_retry to 7(max value), and the problem did not occur again.

Signed-off-by: zhengbing.huang <zhengbing.huang@easystack.cn>
Signed-off-by: Philipp Reisner <philipp.reisner@linbit.com>