drbd: introduce RQ_NET_READY indicating when a sender may process the request
Previously RQ_NET_QUEUED had the meaning that the sender might store
pointers to the request, and that it may process it. Split these 2 roles
into separate flags. Now RQ_NET_QUEUED indicates that the sender might
store pointers to the request, for instance req_next, and RQ_NET_READY
indicates that the sender may process the request.

This is preparation for a change to how a Primary node sends
P_OUT_OF_SYNC packets. It also reduces the complexity of the code.