drbd: fix re-initializing activity log on resize
During online resize, the on-disk offset of internal meta data changes.
We have to move the bitmap, and re-initialize the activity log ring buffer.
That is not an atomic operation, so we
 - mark the bitmap and activity log as "unreliable" in our superblock
 - re-initialize the activity log
 - write the bitmap
 - mark both as reliable again.
In case we crash before the first write of the new superblock,
"nothing has happened yet". If we crash during the other operations,
we know we have to do a full sync.

Since "Optimize AL updates in a two-node cluster",
that is v9.2.8 resp. v9.1.19, __al_write_transaction(),
which actually writes the single activity log transaction,
checks for "activity log disabled",
which is the relevant optimization of that change.
But this is also the "mark AL as unreliable" above,
so the re-initialize step does effectively nothing.

If no further event [*] causes an AL transaction to be written,
we end up with an uninitialized activity log ring buffer on disk,
and the next "attach" will fail ("No usable activity log found").

[*] local promotion + write, or diskless Primary peer + write,
or more than one peers + some write somewhere

The impact of that regression was "accidentally" mitigated with
53fe7073c539 ("drbd: make "day0" bitmap slots reliable again")
drbd-9.2.14: as long as you have more than two bitmap slots,
we must not optimize away the activity log updates.

Drop the "write_al_updates" test in __al_write_transaction().
The path via drbd_al_initialize() does not want to check it,
and the normal path via al_write_transaction() already checks it anyways.

Fixes: 749ec5708d38 ("drbd: Optimize AL updates in a two-node cluster")