drbd: scale received bitmap to bitmap granularity
For compat reasons, we send the bitmap as if bm_block_size was 4k.

We need to scale to our actual bm_block_size.

"compressed", we need to scale the run length.
"plain", we scale the bits.

Still need to deal with "partial bits"
both in the struct bm_xfer_ctx cursor,
and while aggregating set bits.

Because we exchange the bitmap both ways,
the effective bitmaps are equivalent again
after the full bitmap exchange.