drbd: drbd_md_get_buffer: do not give up early
We used to have an unbounded wait_event.

Then we tried to introduce a warning:
3242f5a08 drbd: Warn in case drbd_md_get_buffer() takes longer than 10 seconds
but effectively gave up instead of just warn,
"continue" jumpes not "to the start of the loop", but before while (0)..

Instead of correcting that "in the spirit", this was made explicit:
d27e2d19e drbd: drop useless use of do {} while (0)

Other commits in that area just mangled the message a bit.

The right thing to do is to wait until we get the buffer,
or the backend disk is detached, not just give up after an
arbitrary timeout.