drbd: Allow stand-alone devices to omit bitmap allocation
Introduce a new configuration option 'bitmap' in the disk options
section (d_bitmap in the code). It defaults to yes. When it is set to
no, the DRBD driver does not allocate memory to hold the bitmap.

As a consequence, a resource fails new-peer commands when it has a
single device with bitmap allocation disabled.

A resource also refuses to attach a disk when peers are defined.