docker: improve RPM selection from requires
We compare the "Requires:" section of the DRBD rpms against the currently
running kernel. This works well enough, but sometimes RHEL still manages to
break even this, by making changes to for one reason or another do not show
up in the rpm Provides/Requires section.

This is exactly what happened with RHEL 9.4 kernel 5.14.0-427.88.1. This kernel
is not compatible with previous versions of the 9.4 release, leading to an
immediate kernel panic on loading such a module.

For the above case, we provide an updated kmod, but our logic selecting the RPM
to load failed, because it only looked at the RPM requires/provides. Now, after
filtering based on RPM metadata, sort again by using lbdisttool.py.

Fixes: c4c0e3bb538e ("docker: find matching RPM based on metadata")