drbd: avoid triggering oom-killer when allocating the bitmap
The DRBD bitmap can be very large. We already have mechanisms to handle
bitmap allocation failure. However typically oom-killer was invoked
instead of the allocations failing. This causes all sorts of chaos,
including reboots. If the DRBD devices are brought up on boot, this can
result in a boot loop.

Avoid this by setting __GFP_NORETRY so that oom-killer in not invoked
for the allocations.

Also set __GFP_NOWARN. Our own messages in the kernel log and from the
userspace utilities are sufficient.