drbd: use compound pages to optimize large I/O
Optimize the number of page allocations for large I/O requests.
Instead of allocating a large number of pages, try a "higher order"
allocation and make it a compound page. If that fails,
fall back to PAGE_SIZE-sized (order 0) allocations.

Before these changes, writing in 1MiB chunks: 858 MB/s 947 MB/s 892 MB/s
With these changes for the same test:         954 MB/s 988 MB/s 939 MB/s

On average, from 899 to 960, an improvement in throughput of +6.8%

Of course, this result heavily depends on the environment. The only
point I want to make is that it improves performance measurably when
writing in large chunks. It also decreases CPU utilization on the
secondary.