drbd_transport_tcp: fix potential stack corruption
We use on-stack completions to wait for successful TLS handshakes. In the
event the handshake is not successful, we tear down the sockets using
dtt_socket_free(), which also cancels the completion.

We did not consider that the completions are only declared in the scope of
the "if (tls)" block, but clean up happens one scope above, so strictly
speaking the on-stack completions are no longer valid, but they may be still
in-progress.

To fix this, ensure that the completions are declared in the scope that they
are cancelled in.

Fixes: 3886480eb0be ("drbd_transport_tcp: add kTLS support")