drbd: only send P_PEERS_IN_SYNC for up to 4 MiB when resync finished
Without this fix, we sent P_PEERS_IN_SYNC for evey 4 MiB extent from the
point where the resync stopped until the end of the device. If the
resync was small, then this could be more-or-less the entire device. A
40 TiB device has 10_000_000 such extents. Iterating through them and
sending P_PEERS_IN_SYNC for each caused performance problems.

The typical symptom of this problem was write completion latencies of
several seconds after a small resync.

The problem could also cause the "blocked:upper" status to be shown.
This occurred when some pending bitmap work was stuck behind
w_update_peers() in the queue of the worker thread.

Fixes: bc218ad64640 ("drbd: only send P_PEERS_IN_SYNC every 4MiB")