drbd: call bdev_freeze on user suspend
bdev_freeze ensures that any file systems or other users of a block device can
do some last minute flushing to get into a "consistent" state.

Since we probably get stuck if we try to freeze an already suspend device, we
also need to set a flag on which devices are already suspended.

Included are also the necessary compat patches for building on all relevant
kernels.