drbd: fix an aspect of on-suspended-primary-outdated=force-secondary
Fix the logic for setting fail_io ('force-io-failures' in drbdsetup
status). So far, the code can set it on a secondary or a device with
no openers.

Fix this by:
Setting fail_io only when force_demote is true, and at least one
opener is present. In addition, recheck that the resource is primary
while holding the state_rwlock, and hold the open_release between
checking for openers and setting fail_io.